+++ bugsnag/src/main/java/com/bugsnag/Configuration.java
@@ -8,7 +8,6 @@ import com.bugsnag.callbacks.JavaxServletCallback;
 import com.bugsnag.delivery.AsyncHttpDelivery;
 import com.bugsnag.delivery.Delivery;
 import com.bugsnag.delivery.HttpDelivery;
-import com.bugsnag.delivery.SyncHttpDelivery;
 import com.bugsnag.serialization.DefaultSerializer;
 import com.bugsnag.serialization.Serializer;
 
@@ -28,7 +27,6 @@ import java.util.concurrent.atomic.AtomicBoolean;
 public class Configuration {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(Bugsnag.class);
-
     private static final String HEADER_API_PAYLOAD_VERSION = "Bugsnag-Payload-Version";
     private static final String HEADER_API_KEY = "Bugsnag-Api-Key";
     private static final String HEADER_BUGSNAG_SENT_AT = "Bugsnag-Sent-At";
@@ -36,9 +34,9 @@ public class Configuration {
     public String apiKey;
     public String appType;
     public String appVersion;
-    public Delivery delivery = new AsyncHttpDelivery(SyncHttpDelivery.DEFAULT_NOTIFY_ENDPOINT);
-    public Delivery sessionDelivery =
-            new AsyncHttpDelivery(SyncHttpDelivery.DEFAULT_SESSION_ENDPOINT);
+    public Delivery delivery;
+    public EndpointConfiguration endpointConfiguration;
+    public Delivery sessionDelivery;
     public String[] filters = new String[]{"password", "secret", "Authorization", "Cookie"};
     public String[] ignoreClasses;
     public String[] notifyReleaseStages = null;
@@ -57,6 +55,11 @@ public class Configuration {
         addCallback(new AppCallback(this));
         addCallback(new DeviceCallback());
         DeviceCallback.initializeCache();
+
+        endpointConfiguration = EndpointConfiguration.fromApiKey(apiKey);
+
+        this.delivery = new AsyncHttpDelivery(endpointConfiguration.getNotifyEndpoint());
+        this.sessionDelivery = new AsyncHttpDelivery(endpointConfiguration.getSessionEndpoint());
 
         if (JavaxServletCallback.isAvailable()) {
             addCallback(new JavaxServletCallback());
@@ -120,23 +123,33 @@ public class Configuration {
     }
 
     /**
-     * Set the endpoints to send data to. By default we'll send error reports to
-     * https://notify.bugsnag.com, and sessions to https://sessions.bugsnag.com, but you can
-     * override this if you are using Bugsnag Enterprise to point to your own Bugsnag endpoint.
-     *
+     * @deprecated Use {@link #setEndpoints(EndpointConfiguration)} instead.
+     */
+    @Deprecated
+    public void setEndpoints(String notify, String sessions) throws IllegalArgumentException {
+        setEndpoints(new EndpointConfiguration(notify, sessions));
+    }
+
+    /**
+     * Set the endpoints to send data to. Use this to override the default endpoints
+     * if you are using Bugsnag Enterprise to point to your own Bugsnag endpoint.
+     * <p>
      * Please note that it is recommended that you set both endpoints. If the notify endpoint is
      * missing, an exception will be thrown. If the session endpoint is missing, a warning will be
      * logged and sessions will not be sent automatically.
-     *
+     * <p>
      * Note that if you are setting a custom {@link Delivery}, this method should be called after
      * the custom implementation has been set.
      *
-     * @param notify the notify endpoint
-     * @param sessions the sessions endpoint
-     *
-     * @throws IllegalArgumentException if the notify endpoint is empty or null
+     * @param endpointConfiguration the endpoint configuration
+     * @throws IllegalArgumentException if the endpoint configuration is null or if the notify endpoint is empty or null
      */
-    public void setEndpoints(String notify, String sessions) throws IllegalArgumentException {
+    public void setEndpoints(EndpointConfiguration endpointConfiguration) throws IllegalArgumentException {
+        if (endpointConfiguration == null) {
+            throw new IllegalArgumentException("Endpoint configuration cannot be null.");
+        }
+        String notify = endpointConfiguration.getNotifyEndpoint();
+        String sessions = endpointConfiguration.getSessionEndpoint();
         if (notify == null || notify.isEmpty()) {
             throw new IllegalArgumentException("Notify endpoint cannot be empty or null.");
         } else {